refactor(room): 재생 입장과 joined UI 책임 분리
